requirements.txt
asgiref==3.2.10
certifi==2023.7.22
charset-normalizer==3.2.0
Django==3.1.2
django-crispy-forms==1.9.2
djangorestframework==3.12.2
idna==3.4
Pillow==9.5.0
pytz==2023.3.post1
razorpay==1.3.0
requests==2.31.0
sqlparse==0.4.4
urllib3==1.26.6
setuptools